requirements.txt
asgiref==3.5.2
aspose-words==22.10.0
contourpy==1.0.5
cycler==0.11.0
Django==4.1.2
fire==0.4.0
fonttools==4.37.4
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.6.1
numpy==1.23.4
opencv-python==4.6.0.66
packaging==21.3
pdf2docx==0.5.6
Pillow==9.2.0
PyMuPDF==1.20.2
pyparsing==3.0.9
PyPDF2==2.11.1
python-dateutil==2.8.2
python-docx==0.8.11
six==1.16.0
sqlparse==0.4.3
termcolor==2.0.1
